requirements.txt
pencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
shapely>=2.0.0
open3d>=0.17.0
pyyaml>=6.0